end of day commit - solving the object reference issue
